ran npm init and added node_modules to gitignore
adding node_modules to gitigore ensures that we can't pull node_modules from the repo when it's pulled. This is done because it takes a long time to pull node_modules, a better alternative is to pull and then run "npm install" so that we install node_modules locally as this is much faster.